[Engine] Move String.h -> Utils/String.h
